add if index_exist

+++ db/migrate/20170131125658_remove_index_of_save_by_name.rb
@@ -1,5 +1,5 @@
 class RemoveIndexOfSaveByName < ActiveRecord::Migration[5.0]
   def change
-    remove_index :stories, name: :index_stories_on_user_id_and_saved_at
+    remove_index :stories, name: :index_stories_on_user_id_and_saved_at if index_exists?(:stories, name: :index_stories_on_user_id_and_saved_at)
   end
 end
